refactor(P2P): Moved p2p telemetry metrics into telemetry module
